feat : uuid 생성 api 추가, token 삭제,발급 로직 예외상황 보완 (#36)

+++ src/main/java/Mefo/server/domain/user/service/UserService.java
@@ -102,14 +102,17 @@ public class UserService {
         User user = getLoginUser(email);
         Optional<FirebaseToken> firebaseToken = firebaseRepository.findByUserIdAndDevice(user.getId(), pushAlarmRequest.getDevice());
         boolean pushAlarm;
-        if(firebaseToken.isPresent()){
+        if(firebaseToken.isPresent()&&pushAlarmRequest.getFcmToken()==null){
             firebaseRepository.delete(firebaseToken.get());
             pushAlarm = false;
         }
-        else{
+        else if(firebaseToken.isEmpty()&&pushAlarmRequest.getFcmToken()!=null){
             FirebaseToken token = new FirebaseToken(user, pushAlarmRequest.getDevice(), pushAlarmRequest.getFcmToken());
             firebaseRepository.save(token);
             pushAlarm = true;
+        }
+        else{
+            throw new BusinessException(ErrorCode.TOKEN_NOT_PROVIDED);
         }
         return pushAlarm;
     }
